+++ apps/cowswap-frontend/src/modules/affiliate/hooks/useIsRewardsRowEnabled.ts
@@ -1,0 +1,8 @@
+import { useFeatureFlags } from '@cowprotocol/common-hooks'
+import { isInjectedWidget } from '@cowprotocol/common-utils'
+
+export function useIsRewardsRowEnabled(): boolean {
+  const { isAffiliateProgramEnabled } = useFeatureFlags()
+
+  return isAffiliateProgramEnabled && !isInjectedWidget()
+}