+++ apps/cowswap-frontend/CHANGELOG.md
@@ -1,5 +1,54 @@
 # Changelog
 
+## [2.5.0](https://github.com/cowprotocol/cowswap/compare/cowswap-v2.4.0...cowswap-v2.5.0) (2026-02-25)
+
+
+### ✨ Features
+
+* add xstocks token list ([#7073](https://github.com/cowprotocol/cowswap/issues/7073)) ([f532f26](https://github.com/cowprotocol/cowswap/commit/f532f26aa0d87ee32cac3d5be7ab9ab5f28abb02))
+
+
+### 🐛 Bug Fixes
+
+* prevent input focus when max button is clicked ([#7075](https://github.com/cowprotocol/cowswap/issues/7075)) ([5dccd02](https://github.com/cowprotocol/cowswap/commit/5dccd025b854737547be73144eaf40c1a81bccca))
+* **widget:** apply theme mode from parameters ([6c3125d](https://github.com/cowprotocol/cowswap/commit/6c3125dd09968513e9de3005b5d820d6d1eaa073))
+* **widget:** apply theme mode from parameters ([64f498a](https://github.com/cowprotocol/cowswap/commit/64f498a4ba20e30c79ee62aa5f6a434f92d2a26f))
+
+
+### ♻️ Refactoring
+
+* update quote ID formatting to support ellipsis for long IDs ([#7076](https://github.com/cowprotocol/cowswap/issues/7076)) ([2267fbb](https://github.com/cowprotocol/cowswap/commit/2267fbb917cfdc2ae7292223faac6f5a3ee61212))
+
+
+### 🔧 Miscellaneous
+
+* main -&gt; dev ([9c6611e](https://github.com/cowprotocol/cowswap/commit/9c6611e90a4b23e23f6aaadc513de3c6d7f3b562))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/analytics bumped to 2.3.2
+    * @cowprotocol/balances-and-allowances bumped to 2.3.2
+    * @cowprotocol/common-const bumped to 2.3.2
+    * @cowprotocol/common-hooks bumped to 2.3.2
+    * @cowprotocol/common-utils bumped to 2.3.2
+    * @cowprotocol/core bumped to 2.4.0
+    * @cowprotocol/ens bumped to 2.3.2
+    * @cowprotocol/events bumped to 3.3.2
+    * @cowprotocol/hook-dapp-lib bumped to 2.0.6
+    * @cowprotocol/iframe-transport bumped to 2.0.6
+    * @cowprotocol/multicall bumped to 2.3.2
+    * @cowprotocol/permit-utils bumped to 2.3.2
+    * @cowprotocol/snackbars bumped to 2.0.8
+    * @cowprotocol/tokens bumped to 2.4.0
+    * @cowprotocol/types bumped to 3.3.2
+    * @cowprotocol/ui bumped to 2.3.2
+    * @cowprotocol/wallet bumped to 2.3.2
+    * @cowprotocol/wallet-provider bumped to 2.1.6
+    * @cowprotocol/widget-lib bumped to 1.2.2
+
 ## [2.4.0](https://github.com/cowprotocol/cowswap/compare/cowswap-v2.3.0...cowswap-v2.4.0) (2026-02-24)