+++ libs/balances-and-allowances/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## [2.3.2](https://github.com/cowprotocol/cowswap/compare/balances-and-allowances-v2.3.1...balances-and-allowances-v2.3.2) (2026-02-25)
+
+
+### 🔧 Miscellaneous
+
+* main -&gt; dev ([9c6611e](https://github.com/cowprotocol/cowswap/commit/9c6611e90a4b23e23f6aaadc513de3c6d7f3b562))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/common-const bumped to 2.3.2
+    * @cowprotocol/common-hooks bumped to 2.3.2
+    * @cowprotocol/common-utils bumped to 2.3.2
+    * @cowprotocol/core bumped to 2.4.0
+    * @cowprotocol/multicall bumped to 2.3.2
+    * @cowprotocol/tokens bumped to 2.4.0
+    * @cowprotocol/types bumped to 3.3.2
+    * @cowprotocol/wallet bumped to 2.3.2
+    * @cowprotocol/wallet-provider bumped to 2.1.6
+
 ## [2.3.1](https://github.com/cowprotocol/cowswap/compare/balances-and-allowances-v2.3.0...balances-and-allowances-v2.3.1) (2026-02-24)